move to easconfig
